FEBio3: Moved all classes in FEMesh.h to their own files.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio3@11268 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioFluid/FEFluidResidualVector.cpp
@@ -28,7 +28,7 @@ void FEFluidResidualVector::Assemble(vector<int>& en, vector<int>& elm, vector<d
     //#pragma omp critical
     {
         // assemble the element residual into the global residual
-        int ndof = fe.size();
+        int ndof = (int)fe.size();
         for (i=0; i<ndof; ++i)
         {
             
@@ -47,7 +47,7 @@ void FEFluidResidualVector::Assemble(vector<int>& en, vector<int>& elm, vector<d
         }
         
         
-        int ndn = ndof / en.size();
+        int ndn = ndof / (int)en.size();
         // if there are linear constraints we need to apply them
